Add the ability to mark tasks as done
Items class: keep track of the list of user text inputs

We need a way for the Item class to keep track of the status of the
task.

Lets rename the Items class to be called TaskManager and it's
methods to better reflect the purpose of the class.
Added setTaskDone method to TaskManager class.

I also added a new class called Task to keep track of the task
description and whether if it is done or not.
The Task class has a public method getTaskStatus to return the done
status and description of the task as a string.