Refactor code to comply with coding standards
Classes Duke, Task and TaskManager currently comply with the coding
standards.

This can affect the readability of the code.

Lets update all the classes and public methods to include a javadoc
and ensure that scope of the variables are defined at the smallest
scope possible.
I also ensured that whitespaces are used in statements to improve
readability.