Refactor Task, Todo, Deadline and Event class for readability
There are many single strings that are scattered about in the print
statements, causing code to be hard to read

Lets have named constants for opening and closing of strings to be
printed out. We also have constants for the tick and cross icons.
We also have constants for task tags.