Add handling of exceptions for TaskManager class
TaskManager does not handle invalid or missing user parameters.

Lets throw a MissingTaskLiteralException if the program does not
detect a /at or /by and prompts the user to re enter a new command.

For setting task to be done, we also handled cases where parameter
passed in is not a integer or the integer is out of range of the
available list.