Add delete function to TaskManager
The TaskManager class is unable to delete tasks from the task list.

Lets first convert the array we have into an ArrayList and replace
all the corresponding array methods with ArrayList methods (such as
.get() and .add()).
Next I added a new method, deleteTask(). I also removed exception
when the ArrayList is full.
Lastly, I added an interface to the taskManager so that it knows to
override the public methods for better abstraction.