Add ability to load data from disk
Data is current not saved and information will be gone once the
programm exits.

Lets create a FileManager class that can load data from
./data/duke.txt. The Path type is used to pass to the scanner to
ensure cross platform file paths are compatible.

Once data is retrived, we then pass it to the DataParser class to
convert the String text to Task Objects to be added to the
taskManager.

I have also overloaded the Task constructor so that it can accept
isDone boolean.