Add ability to save task to hard disk
Currently we can only load data from our drive but unable to write
to duke.txt.

Lets create a method under FileManager class to write data to
duke.txt. We also created a FileWritable interface for the task
class for it to convert the object to a string to be saved.