Handle file read format error
The DataParser class might give errors when the format of duke.txt
is not correct.

Lets handle the exception by creating FileCorruptedException to
throw and this is handled by the FileManager. The FileManager will
then delete the corrupted file and then recreate a new duke.txt.

Added a new test case to test when duke.txt is corrupted