Refactor TaskManager class
Previously, the TaskManager class handles converting the parameters
into taskNumber, description and dateTime. This makes the class do
too much.

Lets move the parsing of parameters to ParameterParser class, which
its purpose is to parse the userinput into integers for delete and
done command, and description and datetime for add command.

The ParameterData class stores the data parsed by the
ParameterParser for better readability.

The TaskAction interface and Command classes are updated to follow
this new implementation.