Refactor FileManager class
previously Filemanager class methods are all static.

Lets create a Filemanager class that handles all the loading and
writing of data through an instance of the class.

The code inside the FileManager class are also extracted out to
ensure SLAP principle is followed. This give more readability.

Fix bug due to missing break in switch statement.