Refactor output code to Ui classes
Right now all the printing of user output is everywhere in my code.

Lets create a ui package to handle all the printing of output.
This is splited up into CommandUi, FileUi, InputUi, and TaskUi,
which all inherits the Ui abstract class.

Each class will handle the printing of error and success cases and
printing information.

The InputUi class will also be incharge of reading in the user input
and trimming it.