Add datetime for Duke
Right now we are only passing the datetime parameter of deadline and
event as a string.

Lets teach duke to understand more on date and time and we first
convert all the string parameters of datetime into a LocalDateTime
object along with the other classes that uses the datetime
parameter.