Add a datetime parser class
Now we can store our datetime as LocalDateTime objects but are
no able to parse from our user input yet.

Lets create a DateTimeParser class to handle the user input.
We also have created an enum of all valid datetime format that my
program accepts.
The DateTimeParser will try every single format to see if the user
input fits any of the valid format. If there are no valid format,
it will throw a DateTimeFormatException and this will be handled
by the CommandUi class, informing user that datetime format is
wrong.