Add ability to print datetime
Now we can convert user inputs into datetime objects, we need a way
to display the time back to them.

Lets have a DateTimePrintable interface that the Event and Deadline
class will implement so that they can print out the datetime
pattern based on a formatter.
I have also updated the DateTimeFormat Enum to include more formats
the user inputs we can accept.
When writing the task to file, the time format used is also clean
and easy for the user to change manually.