Add ability to sanitize datetime inputs
Right now user need to type Jan with front capitalized and AM/PM in
all caps in order for the program to register the correct datetime.

Lets sanitize the user inputs such that it will auto lowercase the
input and then proceed to replace it with the correct format.

E.g JAN -> jan -> Jan and Am -> am -> AM