Add ability for list command to filter by date
Lets allow user to be able to filter their todo list by date.
I have added a /on literal to the list command and followed by the
date to query for. This will return a list of all tasks on that
date. If no result is found, an error will be returned to the user.

To simplfy the command further, user can enter list /on today to
get the list of all the tasks on today's date.

All todos will have a date of today.