Update find command
Right now my list command takes in a /on parameter to list by date.
While my find command takes in a user string to search description
by string.

Lets move the /on parameter from the list command to the find
command. So now the find command can either search by keyword or by
date or both.

Lets also add in the ability to search before or after by using the
/bf and /af literal instead of the /on literal. This way they can
search for any tasks before or after the given date.