Update user input interface
The input for user commands ui seems boring.

Lets enhance by adding a header to the output so that it shows
the current number of todos, deadlines and events the user
have in their tasks.

Lets inform users the total number of tasks saved when the user
wants to exit the program.